+++ src/Events/ApprovalExpired.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Enadstack\Approvio\Events;
+
+use Enadstack\Approvio\Models\ApprovalRequest;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class ApprovalExpired
+{
+    use Dispatchable, SerializesModels;
+
+    public function __construct(
+        public ApprovalRequest $request,
+    ) {
+    }
+}